Added centos-7.0 ec2 template
This patch also takes a stab at removing some of the duplication we introduced
while adding debian-7.6 for ec2 by combining efforts with existing scripts.

It also enables proper usage of the puppet-masterless provisioner to make the
build more consistent with the non-ec2 templates. This also means we can avoid
having to do any form of transer to the EC2 instance ourselves, instead we just
let the provisioners do this work as needed. The ec2.pp file has now been removed
and base.pp is now capable of altering the classes based on the provisioner
chosen, so we don't install vmware tools when EC2 is being provisioned for
example.

The bootstrap-puppet.sh script is now capable of installing PE via the NFS mount
or via a tarball if a PE_URL is provided.

The install-puppet-enterprise.sh still overlaps with some of this activity, but
since its used by other elements we don't manage I've avoided changing this for
now.

The debian-7.6 template for EC2 has been modified also to benefit from these
refactors.

Signed-off-by: Ken Barber <ken@bob.sh>