Solaris 11.2: Add release/evaluation package
Solaris now requires a "licence" package to be installed as a dependency of
many base packages; such as the GCC runtime libraries. Installation of this
package requires an `--accept` flag to be passed to PKG, which breaks
automated builds and installations which aren't expecting a licence check.

This patch adds the `release/evaluation` package to the automated install
manifest for Solaris 11.2.