(RE-7417) Update template naming to match repo
As per following comment from Heath:
We should also look at the hierarchy and naming we currently use within
the templates/ directory. The current tooling within the pipeline
expects all json to be located within the template_name dir ( and not
within subdirectories ). It also relies upon a specific naming scheme
for the templates -- $arch.$type.base for the base image,
$arch.$type.vagrant.$config for those that pass through the vagrant
post-processor and $arch.vsphere.$config for templates that utilize the
vsphere post-processor.