(RE-7152) Sysprep/Autounattend for Admin autoulogin
This has proved tricky to achieve, but appears to be necessary in the
Autounattend files in order to obfuscate the Administrator passwords.
The following guide was used to get the approprpriate settings:
http://www.acumenitsupport.com/windows/
unattend-xml-to-sysprep-your-image-and-not-disable-local-administrator

Also setup specific Administrator and puppet users accounts in both
base unattended and generalised unattended. Both are necessary as the
sysprep/generalise phase wipes some of these.