(RE-7663) Change to Simplied Admin Password
This allows us to use a known plaintext password for the initial packer
stages, allowing suitable credentials to be used especially for the
boxstarter restart phase.

The plaintext password will be replaced right at the end of the build
sequence - and may well be done during the cloned machine
initialisation.