(RE-7671) Cleanup Machine for deployment
1. Clear WinSxS (update) space in base image.
2. Use sdelete to zero out unused space.
3. Remove Chocolatey (RE-7671)
4. Delete Temp and other unecc files.
5. Clear Event Logs

Note - the sdelete is only done following the win-update cleanup as this
is the single largest action and gives about a 6G reduction of disk
space. Cleanmgr was evaluated, but only did minor cleanup (<1G so wasn't
considered worth the effort of using for Windows-2012).

The DISM commands used here only work for Windows 2012+, so an
alternative cleanup is required for earlier windows versions.