(RE-7682) Add Chrome Default User Preferences
This allows the blank page (and other Chrome default settings) to be
defined. Also prevent the Chrome First Run dialogues.

Note - there is a bug in Chrome which causes the Chrome first run prompt
to run unless you use special incantations in the master_preferences
file. This edition of the file *seems* to work.

Note also the extra FACTER_xxx facts that need to be added. There may
still be a localisation issue lurking in this w.r.t. Program Files.
O btw - Google installs the 64 bit version into (x86)