(maint) Modify default Windows Explorer UI for Windows templates
Previously the Windows Packer templates were only setting a small number of
Explorer UI settings.  These settings make it a better user experience when
diagnosing issues and also increases the apparent speed of Explorer.  This
commit modifies the Windows Server 2012R2 and 2016 templates to setup Explorer
with the following settings:

- Always show icons, never thumbnails
- Always show menus
- Display file icon on thumbnails
- Display the full path in the title bar
- Show hidden files, folders, and drives
- Show empty drives
- Show extensions for known file types
- Show folder merge conflicts
- Show protected operating system files
- Combine taskbar when full
- Show Administrative Tools on Start Menu
- Do not drag full windows
- Disable font smoothing
- Disable Aero Peek

This commit also brings the 2016 template more into line with the 2012 R2
template.