(IMAGES-76) Change Post-Clone to sysprep operation
The Run Once key in the registry is armed to execute a sysprep command
which runs the sysprep and then the post-clone configuration script.

Add unattended file for all current OS along with an update to the
vsphere json file to upload the unattended file

Removed HKCU registry key configuration to since Default User keys will
now be used to create the Administrator Profile after the sysprep.
Needed to add one registry fix to the post-clone-config to correct the
background screen setting.