(IMAGES-366) Add VirtualBox support to packer
This adds new files necessary to build virtualbox images with packer
using the work already done to build vmpooler images. It follows the
convention of adding new files for new builders instead of having one
parameterized build file.

The autounattend files were modified because the disk partition structure was
incompatible with the expectatons of the virtalbox builder. It was also
modified to remove setting the RemoteSigned executionpolicy for
PowerShell since this template is for Windows2012R2 and that setting is
already present.

The oracle cert was included to install the VirtualBox tools silently.

A new windows-2012r2/files/{{build_name}}.package.ps1 for virtualbox was
added as per convention, even though in this commit nothing was changed
from the vmpooler script.