(IMAGES-101) Add Win-2008R2 Win Update Cleanup
Use the KB2852386 hotfix for cleanmgr to clean the windows udpate
(winsxs) area.

Also use cleanmgr across all OS's for a more comprehensive disk tidy up.
A restart is required after cleanmgr, so the sdelete operation is moved
to a separate script

MAINT commit to be added to update other OS package files to install
desktop-experience