(IMAGES-22) Address DISM issues in Win-10 template
Use Slipstream image to prepare ISO with the latest CU available. This
reduces the overall install time, but may need to be updated prior to
each release to include the latest CU.

The DISM cleanup is done as part of the slipstream process so the
clean-dism script is not used for Windows-10 base build - at least until
bare-metal hardware is introduced as otherwise the jenkins-imaging
builds fail fairly consistently.