(maint) fedora 25 template fixups
add an output_directory variable that can be set with an environment
variable.  this allows us to change the base for the output and if not
set, it will default to the current directory of the packer run. also
remove templates that we're not currently building.

(maint) use open-vm-tools
(maint) f25 network fixups