(MAINT) Insert Vagrant Specific SSH keys.
The vagrant logon for ssh requires its own ssh key. To accomodate this,
add authorized_keys.{vagrant|vmpooler} and move these to the Packer/Init
directory which will copy both as part of the machine arming.
Update all the vsphere json files so as not to load the authorised_keys
onto the A: drive as they are no longer needed.