(IMAGES-416) Fix Vagrant ssh behaviour
Revise the usernames used for the build process so that
"vagrant/vagrant" is used throughout including the
boxstarter-initialisation.

Make sure the Administrator account profile is also created, with an
empty directory for the Administrator Home dir on ssh.

Finally run sysprep to clear the profiles and pick up the default user
settings provided by puppet.