(IMAGES-431) Use Git for Windows instead of cygwin
Replace the git supplied by Cygwin with Git for Windows
The git package is removed from the Cygwin package during installation,
so git commands within the cygwin shell are actuall gitforwin.

Minimal gitforwin install options are chosen using an INF file at
installation time. The INF file is used by all the windows packer
builds, so a global change is made to all the .json files.

Ensure that the credentials manager provided with gfw is disabled and
that github.com is added as a known_host to support
"git@github.com:REPO" type clone operations.

Finally direct the editor to the Cygwin vi to avoid any issues with the
editor for git config or git commit commands.