(IMAGES-386) Resolve Win-2012 build hang
The Windows-2012 build hangs indefinately during the Windows Update.
The first step to address this was to create a formal slipstream build
process for Win-2012 (as the first slipstream was built manually) and to
ensure that all the correct Windows Servicing Stack updates were
installed correctly to allow it to operate with WSUS.

It was also necessary to filter out a number of updates during the DISM
offline servicing phase as these appeared to corrupt the image. Three of
the updates are critical however, to the actual build, so these are
inserted into the main build package script just prior to the windows
update phase.