(IMAGES-386) Hardwire WSUS server for main build
The main build appears to operate when pointing to the WSUS server
rather than Microsoft Update. At the moment, the WSUS server is a dev
machine hosted on Slice to support the Slipstream generation. It is
intendended to replace this with a production server with a known DNS
name.

Until this is done, a static IP address is being used here to get this
build operational.