(IMAGES-481) Add Consistent Notes for VM
Setting a consistent Annotation field with the build date and SHA for
all machines which is displayed in vSphere data. This update is for
windows machines only, so IMAGES-481 is being left as a ticket on the
backlog to ensure the update is applied to all the Linux boxen at a
later date.