(IMAGES-39) Prepare Slipstream ISO for Win-2008
The Slipstream had to be prepared manually in a few stages due to the
necessity of adding Powershell, .Net 3.5 SP1 and .Net 4.5.2 before
boxstarter could be used to collect the updates. Further DISM
is not available yet for Win-2008, so the actual .iso build and
offline servicing had to be done on a Windows-2012R2 machine.

The PS script used on the Win-2012R2 machine is included for reference.