(IMAGES-39) Windows 2008 Template files
This is the set of new files for the Win-2008 Packer Template. It does
not include modifications to the generic windows scripts as these will
be documented in seperate commits.

Windows-2008 doesn't boot using UEFI on the vSphere cluster, possibly
due to not being the latest ESX version (it works correctly on
Workstation and Fusion), so BIOS mode is used instead which means the
paritions need to be revised in the Autounattend.xml script.

The base build assumes that Powershell has already been injected into
the base image using the slipstream process as boxstarter will not
function unless this is present.

All Powershell scripts need to be run elevated otherwise certain
operations don't run correctly (e.g. Puppet doesn't download or install
modules). This is achieved using the Elevated options in the powershell
Packer Provider.

An additional shutdown powershell script is also needed as the packer
shutdown also fails to run correctly leading to incorrect (failed)
termination of the build process. This is probably due to elevation/UAC
issues mentioned earlier.