(IMAGES-39) Fix PS2/Win-2008 Powershell issue
This delight was obtained from: http://www.leeholmes.com/blog/2008/07/30
/workaround-the-os-handles-position-is-not-what-filestream-expected/

Windows-2008/PS2 appears to have a bug which causes scripts to fail
unless this "compatibility" code is run prior to any executable
statement.

It has been added to the windows-env.ps1 script as it is needed in
almost all of the Powershell scripts during the build process.
However, the arm host script is designed to be run on the actual VM to
aid debugging/testing, so the code has been duplicated at the start of
this script (as A:\windows-env isn't assumed to be present).