(IMAGES-98) Revise Process Execution calls
The Cygwin Setup command works async in win-2008 (it may do on other OS
version but the sleep command hides this), so the Start-Process -Wait
cmdlet must be used to run Cygwin install.

The opportunity was taken to standardise on this command for all the
windows installs (IMAGES-96).

The Windows Install script was also refactored to remove copy/pasta and
tidy the code, especially for the sysinternals code.

Use Splatting for the common "Start-Process" arguments.

Note - there appears to be a bug with the Start-Process -Wait command in
PS2 (especially for Win-2012) which caused the 7zip process to fail it
it exited "too quickly", hence the need for the WaitForExit() method in
the 7zip/sysinternals install.