(IMAGES-501) Cleanup puppetres.dll
This is a follow-on from PA-768 where puppetres.dll and associated
registry keys are left on the system following puppet removal. This need
to be cleaned from the vmpooler created images so as emulate a clean
system where Puppet has not been installed.

Delete the HKML\Software Puppet keys for good measure.