(IMAGES-546) Pin to earlier choco/boxstarter
Need to pin to Boxstarter Version 2.8.29 for the Windows-2008 build.
(this isn't actually formally supported by boxstarter, but we were able
to get it working sufficiently under this version). A recent boxstarter
update has broken the build again, so employing "pinning" for this build
only.

WSUS also needs to be disabled for Win-2008 as it is significantly
slower for the update process if this isn't done.