(IMAGES-446) Add Slipstream for Windows 2008r2
Add a formal slipstream build for Windows 2008r2 (prior slipstream was
handcrafted and not based on the Cumulative Update which simplifies the
overall generation).

Note - the DISM tools available on Win-2008r2 aren't mature enough to
actually build the updated slipstream, so offline processing is still
required on a Win-2012r2 machine to actually build the ISO. So this
procedure here just gathers the CAB files.