(IMAGES-578) Update to latest .Net
Use a common .Net install helper function to install the latest .Net
appropriate to the platform (4.6 for Win-2008 and 4.7 for everything
else).

The installation is also moved to start-boxstarter prior to the
chocolately install as this avoids a potential dual install of .Net.
Adding .Net 4.7 also reduces the Windows update footprint.

Update slipstreams for Win-7/2008r2/2012 to make sure a base patch
(KB4019990) is already applied. This also overcomes a hang with recent
windows updates that exceeded the winrm timeout.

Some additional fixes were also needed to the vsphere stage (extra
reboot) to ensure cygwin installs correctly.