(IMAGES-613) Common Base Packer file
Convert all the base builds for Windows VMWare and Virtual Box to use a
common base build file with a variable definition file in each
directory.

Also move the oracle-cert file into the common file area as this is
identical in all virtual box builds.

One other change to note here is the first "post-winrm" command prints
out the log from Start-PSWindowsUpdate.ps1 to the console so that it is
visible in the packer (and Jenkins) output. This is very useful to allow
checking that windows updates and other scripts have run clean during
this phase.