(IMAGES-613) Common Stage 2 files
Create a common vshphere/cygwin and virtualbox/vagrant 2nd build stage
file and update the variable files with any variables needed for this.
Remove the individual Puppet files and replace with a common win-site.pp
file for all builds.

The vagrantcloud.yaml files were only being used for versioning - this
can be obtained from the vars file, so remove as well.