(IMAGES-613) Common Unattend.xml files
Replace all of the individual Unattend.xml/Post-Clone-Autounattend.xml
with common template files and an XSLT transform that inserts the
platform specific parts.

Note the dual Unattend (Base, Generalise) has been merged into a single
file as the sysprep is not longer needed at initial image creation.

Key changes in this PR are:
1. Create
       new file:   templates/win-common/files/AutoUnattendTemplate.xml
       new file:   templates/win-common/files/Autounattend.xslt
       new file:   templates/win-common/files/PostCloneTemplate.xml
       new file:   templates/win-common/gen-xslt.sh
2. Modify .gitconfig so that tmp/ files are not checked in.
3. Update variable files to add platform specific Autounattend values,
   e.g. Image Name, Product Key
4. Delete platform specific autounattend files for each win platform.