(IMAGES-574) PSWindowsUpdate replace Boxstarter 1
Replace Boxstarter with PSWindowsupdate. This is an extract from the
original PoC PR #207.

A new script start-pswindowsupdate is used to replace start-boxstarter.
Note - since chocolatey is no longer required its not installed - this
simplifies the cleanup as the vmpooler images must not have it
installed.

An alternative test/reboot cycle has been implemented to allow all
windows updates to be installed before WinRM is started to allow the
remainder of the packer build to proceed. This cycle has been moved to
the start-pswindows update script which reduces the size/complexity of
the package script. The package script now focuses on:
1. OS Specific changes (e.g. network card)
2. Patches required before the main windows update cycle progresses.