(IMAGES-627) Add Windows-10 Creators Edition & Pro
This is the latest edition of the Windows Build.
Retain the previous (Anniversary Edition) as this is still supported for
business.

The Anniversary edition is also retained as win-10-1607

Also add the Win-10-pro build (IMAGES-575)

Some fixes were needed to the AutoUnattend to Support Windows-10,
especially the need to turn LUA off at the very beginning, and
post-clone for virtual box.