(MAINT) Don't start sshd during post-clone
Small maint change to improve the acceptance reliability.
Starting sshd before we reboot to rename the host causes failures in the
vcloud/beaker hypervisor - suspect this is because it uses IP addresses
rather than the given hostname to detect when the host is alive, thus
breaking the test initialisation sequence. The vmpooler initialisation
is more tolerant of this as it the ssh test uses the given hostname
which isn't active until after the reboot.

Suspect we can further rationalise and tidy the overal ssh startup
sequence (i.e. use Automatic start), but lets just do this for now.