(IMAGES-707) Set SSHD as Automatic in post-clone
The SSHD startup was still happening in a seperate script to the
post-clone operation. This seems to cause a failure for sshd to come up
in the image accpeptance test env.

Simplify this by setting sshd as Automatic as one of the last commands
before the computer is renamed and take the startup out of the
startup-script.
The mkpasswd and mkgroup commands are also copied to the post-clone
operation to ensure these are in place before the first sshd
initialisation.