(WIN-74) Add WMF2/3 Build definitions
Add both the WMF 2.0 and WMF 3.0 build. This first commit provides the
main outline of the work, but some further commits were needed to
address idiosyncrasies associated with PS2 and PS3.

1. Add the wmf2/wmf3 variable definition files/dirs.
2. Scripts to downgrade to WMF2 and add WMF3 if needed.
3. First attempt at getting appveyor 3rd party scripts working.
4. Add gce-startup script (see below).

The gce-startup.ps1 script is run as a scheduled task at boot under the
"appveyor" account as this was found to be the only way that the WinRM
permissions and quotas could be properly granted to allow the packer
scripts to run correctly for both PS2 & PS3.

The wmf2/wmf3 selection is a little kludgy given the restrictions of the
packer format, i.e. you need to put the conditionals in the actual
powershell script - this leads to a redundant reboot for WMF2, but the
time take for this is neglible compared to the overall build time.