(WIN-106) Apply Windows Updates
The PSWindows Update script doesn't work for PS2/3 under WinRM so
remodel the gce-startup scheduled task to perform the windows update.
This uses a model similar to that for the VMWare/Virtualbox images where
the multiple reboots are done while WinRM is off using a very long
timeout on the windows restart.

All redirection to log files are removed from the startup script as the
scheduled task is run using CMD with its output redirected to a log.
This necessitates a reappearance of that lovely bit of refletion for the
PS2 operation.

Note - the windows-update.ps1 script wasn't needed as this was
integrated into the gce-startup script with common functions in the
windows-env.ps1 script