(MAINT) Repoint all windows-env to \Packer\Scripts
Change all references to the A:\ drive to use the C:\Packer tree except
for the initial bootstrap (start_pswindowsupdate) - one of its first
actions is to copy the windows-env.ps1 file to the scripts area.

All the config files are now placed in C:\Packer\Config, so some of the
installation provisioners also need updated to refer to this.