(IMAGES-784) Create Win-2016 HyperV Image
Create an image with HyperV enabled. This also requires nested
virtualisation enabled at the VMWare level, so need seperate
base/vsphere .json files to set the extra vmx data parameters.

Unfortunately there is a bug in OVFTool 4.2.0 which doesn't allow the
export of the "vhv.enabled=TRUE" flag, so a manual post-processing
action will be needed for this template on vSphere.