(PE-23924) Fixes for CentOS/Scientific Linux 6 repo configurations
The CentOS and Scientific Linux mirror repos do *not* include updates in
their base 'os' repos as was previously assumed. This configures yum to
use separate repos for each of the required os/updates/extras/etc repos,
which differ depending on the distro. I've deliberately kept the
configuration branches separate and self-contained for readability,
given how confusing this has been.