(IMAGES-808) Move ssh enablement for VRO 7.3 to the boot command
Previously, ssh had to be manually enabled for the VRO 7.3 x86_64
platform before building the Packer template because Packer required
an ssh connection in order to provision the VM. This is an annoying
and unnecessary manual step.

This commit also adds a fix to the provisioning script, uses the
QA_ROOT_PASSWD_PLAIN for the ssh password (because the changes in
this commit were causing some build failures) and slightly
reorders some code in the provisioning script.