(IMAGES-799) VMWare Sysprep init during build
Move the Initial Sysprep operation to the last stage of the VMWare build
by means of a specialised shutdown (same as the sysprep/restart used for
virtualbox/vagrant).

This should hopefully stabilise the sysprep for the Windows-10 builds
without impacting same for the earlier versions.

Further commits will be needed to generalise this for earlier Windows
versions.

The previous arming methods (including registry entry script) is no
longer needed.