(IMAGES-800) Stop BGInfo console window popup
Powershell when run under task scheduler has an irritating feature of
popping up a window (which grabs focus) every time bginfo runs.

Apparently the only way to fix this is to encapsulate the Powershell
script in a vbs script and run that from the task scheduler.

Apologies for a last millenium fix for this, but it does appear to work.