(IMAGES-31) Add convenient shortcuts to taskbar
Adding:
1. Explorer
2. Powershell
3. Notepad++
4. Chrome
Useful utils to have immediately available especially when windows seems
to make a career of hiding them from you.

NB - this is a "first attempt" at this, which is only partially
successfull for win-10/win-2016, so leaving the IMAGES-31 ticket open
for further work at a later stage.

It is useful though that the powershell button is appearing in the
win-10 taskbar at least