(IMAGES-838) Set Beakerhost for Win-2016
Beakerhostgenerator needs updated with "new" strings for Windows-2019.
So for the purpose of running the packer acceptance tests only,
temporarily setting the beakerhost name to for windows2016.

Note - this is one of the topics on the AI insurgency - the difficulty
of adding new host types with the multiple locations that host
definitions have to be added to - sigh.