(IMAGES-846) Re-direct file cleanup to log file
This output is far too verbose for the main jenkins logs, so redirect to
log files (as it may be of use debugging imaging issues).

Take the opportunity to tidy up the ForceFullyDelete-Path function
pipeline.

Also add in disk optimization immediately prior to the sdelete phase as
a once only operation for final disk preparation.