(MAINT) Variable File cleanup
Move as many variables into the common vars file and cleanup usage
consistency across all files.

Remove unused variables and make indentation consistent across all
builders. This means there are lot of whitespace only changes.

Drop the use of the "env" to load template variables with build values
(e.g. PACKER_SHA). These are now fed in directly using the -var option
in the build command, which will force the use of the platform-ci-utils
gem (which preloads these values). To facilate this, the packer
lowercase variable names now match their previous ENV names (so that the
two can co-exist peacefully for a period until both the Linux and
Windows builds are standardised).

The qa_passwd variable is in fact the plain one, so change this across
the board as well, including in the downstream powershell scripts, so it
is clear that this is not the same as the hashed password used in the
linux builds.