(IMAGES-818) Pause between puppet module installs
For some reason, puppet module install fails to connect to the forgeapi
after installing the first module when building the SLES 15 image. If we
retry the installation after a brief pause, it succeeds. This commit
adds a brief pause between module installs to the bootstrap-aio script. We
need this to build the SLES 15 vmware.base image from the SLES 15 beta
image.

It is worth checking to see if this is still an issue once we switch
over to the SLES 15 GA image. For now, we only add it here to get
something that is buildable. This should not significantly slow down
our other Packer builds, since we really only install 2 or 3 modules
at a time.