(IMAGES-874) Fix Boot command error
Set the boot wait time to 10m (i.e. almost non-existant) to ensure the
first "key-presses" are passed to the console for the "Please press any
key". If we don't get this in early enough, then the EFI boot screen
comes up along with a world of pain.

Also - simplify the boot command string to spacebar/wait/enter sequence,
since the remainder of the windows install is directed from the
autounattend.xml script. The <enter> is needed for win-2012r2 boots.