(IMAGES-875) Revise ISO download handling
Change from a single iso_url, to a list (iso_urls), where the file:///
location is listed first followed by the artifactory location. The url's
are constructed using the base_iso name which allows support for
building using the new vsphere-iso method and the original nested
virtualisation method (which is still needed for virtualbox).

The search order is /srv/packer/iso followed by attempting to download
from artifactory. /srv/packer/iso is also shared as a datastore on
vcenter thus saving space on this volume as if the file exists in the
.iso area, a copy is not needed in the packer_cache.