(IMAGES-880) Fixup iso_names in platform specs
Fix the individual iso_names in the platform builds.

The vsphere guest name is also added to each as appropropriate.
We need different vmware guest names for vcenter as the version of
vmware fusion is well out of date and doesn't "support" windows-10/2016
guest names. The vcenter/vmware supported list of guest names is
available at:
https://pubs.vmware.com/vsphere-6-5/index.jsp?topic=%2Fcom.vmware.wssdk.apiref.doc%2Fvim.vm.GuestOsDescriptor.GuestOsIdentifier.html