(IMAGES-890) Fix PE Driver Paths for 2008/7 builds
Windows builds earlier than Win-8 (i.e. Win-2008, 2008R2 and 7) need to
use Vista Driver paths for the initial PE phase on vsphere-iso so that
the pvcsci drivers are loaded and the hard drive is visible.

Use similar method to what is used for Platform/firmware selection to
make this choice.

DriverPaths are only needed for the initial autounattend file and not
the post-clone one.