(IMAGES-907) Set Firmware to BIOS for vcenter/vmware builds
The recent ESXI 6.7 upgrade seems to have produced a side effect
where the Firmware setting is ignored when creating the machine,
i.e. we are unable to set it efi - so force the Firmware to "bios"
for the vmware builds only (no changes to virtualbox).

This is a hacky change both in having to force it to bios and
using gen-xslt to do the job (instead of having platform-ci-utils
to the json handling as per IMAGES-888)