(IMAGES-916) Fix Path Delete Function
Prior to this change, the ForcefullyDelete-Path function did not have
a `process` block. This caused a bug such that only the last directory
in a list given to the function via the pipeline was actually processed
and deleted. This change ensures that the full collection of directories
piped to the function are processed and deleted.

Manual testing was done to ensure that any errors encountered deleting
a directory would not cause the function to stop processing the
remaining items in a given collection.