(IMAGES-917) Seperate the bootstrap stage from win update
If a failure happens in the bootstrap phase, it won't get notified until
after a very long timeout period (several hours). So minimise the bootstrap
phase to just installing the hypervisor tools/drivers and any other minimal
network operations so that WinRM can be enabled and control passed back to
packer.

The Windows update phase is initiated later in a seperate phase using the
technique successfully used with the Appveyor build, where WinRm is disabled
just before a packer-reboot provider to allow a further reboot cycle. This
cycle is ended by re-enabling WinRM - thus allowing Windows Update to use as
many reboots as needed.

Also - have changed the startup/reboot sequence to use a Scheduled Task on
Startup to continue the sequence across multiple reboots. This has the
added advantage of being able to run the script elevated, thus getting
round the restriction that WSUS cannot be used across WinRM sessions.

It has also allowed the removal of the $PROFILE Core edition workaround.

This gives a more immediate experience to the jenkins log that the boot sequence
has completed at least.