(IMAGES-918) Localise Admin User/Password
Issue with French/German localised Admin names for the scheduled Task work.
Fix so that we can pass the correct Username/Password into the scripts.

This initial commit allows the bootstrap-packer command line to be modified
to use the localised Administrator Name and Password which is needed to prime
the scheduled task.

It also helps resolve an apparent issue with the win-2019 initial build.