(IMAGES-909) Set Firmware to EFI for vcenter builds
Point the builder to use the new "puppet" build of the
packer-build-vsphere-iso plugin (which is deliberately
named as packer-build-vsphere-plugin-iso to disinguish that
we are using a custom puppet built one until such time as
jetbrains release a revised plugin.

Having the new plugin in place means that we can now direct
the firmware for vcenter builds to efi (except for win-2008
of course). The firmware setting is also removed from the
vmx parameters, where it appeared to cause problems with
ovftool and esxi 6.7