(IMAGES-974) WindowsUpdate Tests
Add Smoke tests to verify that WindowsUpdate has happened including
1. Updates have been installed in the last 12 hours.
2. There are no pending updates.
3. There are no pending reboots

This commit also resolves a serious issue with the windows update
during the shutdown sequence (Invoke-Reboot) which allowed a second
Get-WuInstall to run and interfere with the shutdown sequence.

Some cosmetic changes were also made to the Windows Update scripts
associated with this work as well as minor fixes to the Windows Core
builders for the Server Core type.