(IMAGES-985) Using Pswindowsupdate 2.0.0.4
Updating to Pswindowsupdate 2.0.0.4 (latest version) for Windows
10/2016 only. This appears to work slightly better for these
versions. However, it doesn't play nicely with earlier versions
of windows, so stick to Version 1.6.1.1 for the remainder.

PS3+ auto-imports modules if the PsModulePath is set correctly, and
in any case PsWindowsUpdate 2.0.0.4 requires this to be set correctly
as it uses a .dll which throws a reference error if not set.

This commit also has a few minor fixes for any remaining Windows
ReleaseID parameters in the vars files for the platform version checks.

This commit also contains a fix for IMAGES-984 - allow a more generous
wait at system boot before starting the Windows Update task.