(IMAGES-840) Add WinRm over HTTPS
Enable HTTPS Winrm access during the vmpooler startup operation
where we know what the machines DNS name is going to be. This is
post machine rename as the earlier versions of window seemed to
be fussier about this.

Win-7/2008/2008r2 also require the makecert utility so adopt this
across all versions. Makecert since its a small .exe is added
directly to the repo.