(IMAGES-851) Stop Notepad++ update check
Disable the Notepad++ updater during the installation to suppress
the irritating update prompt when running the editor.
Version 7.5.2+ of npp conveniently provides an install option
(/noUpdater) for this.

Also take the opportunity to puppetise the installation of this package
and update the version to 7.6.2

The windows package install tests need to be moved to the "puppet"
test stage as the Notepad++ is no longer installed by the
"install-win-packages.ps1" provisioner. This provisioner will be
removed once the remainder of the windows packages are puppetised.
(see IMAGES-636)