(IMAGES-1005) Force Firmware to be set
Previous firmware defaults no longer work, so revise to require the
firmware to be set explicitly rather than default to efi - this allows
bios or efi builds to be made, and those that need to default to bios
(windows 2008) to stay at bios.

Add default to the main vcenter build to ensure it works until
platform-ci-utils is fixed up.