(IMAGES-1005) Platform9 Post Clone/Sysprep
Setup the post-clone scripts for Platform9 and perform the sysprep.

VMWare tools is also be removed in this stage. Since the removal
breaks WinRm, a RunOnce key is used to reboot the machine and
run the init-sysprep session again. Otherwise the script hangs
once the tools are removed (as the network drivers close the network
connection).