(IMAGES-1005) Install Platform9 Drivers (#418)
* (MAINT) Remove uneeded files

Remove Start-Profile globally as it hasn't been needed for a while.
Also fixup error in shutdown command parameters

* (IMAGES-1005) Force Firmware to be set

Previous firmware defaults no longer work, so revise to require the
firmware to be set explicitly rather than default to efi - this allows
bios or efi builds to be made, and those that need to default to bios
(windows 2008) to stay at bios.

Add default to the main vcenter build to ensure it works until
platform-ci-utils is fixed up.

* (IMAGES-1005) Install Platform9 Drivers

Download and install device drivers and platform 9 agent packages.

* (IMAGES-1005) Platform9 Post Clone/Sysprep

Setup the post-clone scripts for Platform9 and perform the sysprep.

VMWare tools is also be removed in this stage. Since the removal
breaks WinRm, a RunOnce key is used to reboot the machine and
run the init-sysprep session again. Otherwise the script hangs
once the tools are removed (as the network drivers close the network
connection).