(MAINT) Allow Mock build to Fail Gracefully
Use the new packer array variables to vary the valid_exit codes from the
default fail for acceptance tests for the Mock build. This allows the
windows update validation (and other Pester tests) to fail gracefully
without aborting the build.

The Mock build is a very useful build for pipeline througput testing
with a shortened build cycle.

Also fixup some minor changes with the vars.json file as the original
mock was based on the Insider Preview which is no longer valid.