(MAINT) Remove Redundant ErrorActionPreference
This was noted when investigating the lack of Puppet Error Config
checking. Although this statement was at the top of almost every
powershell script, it was effectively redundant as one of the first
statements in windows_env.sh reversed the preference to "Continue".

So doing a MAINT commit here to remove these redundant lines.