(IMAGES-1018) Fix Puppet Configure Errors
Revise the puppet Configure stage to run as a scheduled task so that
need for rebooting is properly deteched and acted on.

Within this, the configure loop to properly trap and address Puppet
configuration errors - to date these have been logged but not
correctly actioned.

The win-2008 json file needed some additional remedial work to bring it
up to the same spec as the main builds.