(IMAGES-1018) Puppet Errors aren't checked correctly (#435)
* (MAINT) Remove Redundant ErrorActionPreference

This was noted when investigating the lack of Puppet Error Config
checking. Although this statement was at the top of almost every
powershell script, it was effectively redundant as one of the first
statements in windows_env.sh reversed the preference to "Continue".

So doing a MAINT commit here to remove these redundant lines.

* (MAINT) Log Error output for sched tasks

stderr wasn't being logged to the respective log files for various
reboot type cycles within the packer build (e.g. bootstrap phase).
Correct this as a minor maint issue ahead of the puppet configure
restructure.

* (IMAGES-1018) Fix Puppet Configure Errors

Revise the puppet Configure stage to run as a scheduled task so that
need for rebooting is properly deteched and acted on.

Within this, the configure loop to properly trap and address Puppet
configuration errors - to date these have been logged but not
correctly actioned.

The win-2008 json file needed some additional remedial work to bring it
up to the same spec as the main builds.

* (IMAGES-1018) Use Hiera for Build Parameters

To date, build parameters to Puppet have been passed in using env
variables. Change this to use a hiera data file that is generated
pre-build using the internal packer variables.

This is likely to be extended to pick up the pester test parameters and
other parameters.